Generalize smart-routing engine (rename from intelligent routing)
Refactor the Codex smart-routing internals into a harness-agnostic core so a
second agent can reuse them, and rename the feature "intelligent routing" ->
"smart routing" throughout (package dir smart_routing/, symbols, CLI flags
--enable/--disable-smart-routing, hidden codex-router-hook, user-facing
strings) per naming guidance.

- smart_routing/routing.py + hooks.py: shared routes:select call, decision
  shape, canary/audit bookkeeping, and hook add/remove machinery.
- smart_routing/codex_routing.py + codex_hooks.py delegate to the shared core
  (behavior unchanged).
- State key is the agent-neutral "smart_routing_enabled" so a later agent can
  share one opt-in.
