--- src/ucode/intelligent_routing/__init__.py
@@ -1,1 +1,0 @@
-"""Intelligent-routing integrations for supported coding agents."""
