+++ src/ucode/smart_routing/__init__.py
@@ -1,0 +1,1 @@
+"""Smart-routing integrations for supported coding agents."""
